+++ include/atomic/mn_atomic_flags.hpp
@@ -75,6 +75,7 @@ namespace mn {
     }
 
 
+
 }
 
 #endif