+++ include/atomic/mn_atomic_primary_types.hpp
@@ -5,18 +5,18 @@
 *This file is part of the Mini Thread Library (https://github.com/RoseLeBlood/MiniThread ).
 *Copyright (c) 2021 Amber-Sophia Schroeck
 *
-*The Mini Thread Library is free software; you can redistribute it and/or modify  
-*it under the terms of the GNU Lesser General Public License as published by  
+*The Mini Thread Library is free software; you can redistribute it and/or modify
+*it under the terms of the GNU Lesser General Public License as published by
 *the Free Software Foundation, version 3, or (at your option) any later version.
 
-*The Mini Thread Library is distributed in the hope that it will be useful, but 
-*WITHOUT ANY WARRANTY; without even the implied warranty of 
-*MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU 
+*The Mini Thread Library is distributed in the hope that it will be useful, but
+*WITHOUT ANY WARRANTY; without even the implied warranty of
+*MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
 *General Public License for more details.
 *
 *You should have received a copy of the GNU Lesser General Public
 *License along with the Mini Thread  Library; if not, see
-*<https://www.gnu.org/licenses/>;.  
+*<https://www.gnu.org/licenses/>;.
 */
 #ifndef _MINLIB_5ac13f75_2f13_4bb7_acae_7b8f27f78e20_H_
 #define _MINLIB_5ac13f75_2f13_4bb7_acae_7b8f27f78e20_H_
@@ -33,12 +33,13 @@
 #error Unsupported compiler / system.
 #endif
 
+
 namespace mn {
      /**
      * @brief Only for declar special version for types
      */
     template <typename T>
-    struct _atomic : BASE_ATOMIC_TYPE<T> { 
+    struct _atomic : BASE_ATOMIC_TYPE<T> {
         using base_type = BASE_ATOMIC_TYPE<T>;
         using self_type = _atomic<T>;
         using value_type = typename base_type::value_type;