+++ include/miniThread.hpp
@@ -53,7 +53,6 @@
 #include "queue/mn_workqueue.hpp"
 
 #include "mn_ringbuffer.hpp"
-#include "memory/mn_mempool.hpp"
 #include "mn_shared.hpp"
 
 #include "mn_atomic.hpp"